# ajk/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Emergency Alert API",
        "description": "Emergency Alert API documentation",
        "termsOfService": "",
        "contact": {
            "name": "",
            "url": "",
            "email": ""
        },
        "license": {
            "name": "unlicensed",
            "url": ""
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/authorities": {
            "get": {
                "tags": [
                    "authority-resource"
                ],
                "operationId": "getAllAuthorities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Authority"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "authority-resource"
                ],
                "operationId": "createAuthority",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Authority"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Authority"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/authorities/{id}": {
            "get": {
                "tags": [
                    "authority-resource"
                ],
                "operationId": "getAuthority",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Authority"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "authority-resource"
                ],
                "operationId": "deleteAuthority",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Authority": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string"
                    }
                }
            },
...
        }
    }
}